requirements.txt
# Discord
discord.py[voice]>=2.3.2
py-cord>=2.4.1

# YouTube
yt-dlp>=2023.12.30
google-api-python-client>=2.108.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Audio Processing
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.1
asyncio>=3.4.3

# Data Handling
pydantic>=2.5.2

# Logging
colorlog>=6.8.0
